# doc/model/swagger-2.0.yaml
swagger: '2.0'
info:
  description: REST API for the 5GCity's SDK
  version: '1.0'
  title: '5GCity SDK REST API '
  contact:
    name: Nextworks S.r.l.
    url: 'http://www.nextworks.it'
    email: info@nextworks.it
  license:
    name: Apache License Version 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0"'
host: 'localhost:8081'
basePath: /
...
paths:
...
  /sdk/sliceManagement/slices:
    get:
      tags:
        - slice-controller
      summary: Get the complete list of slices available in database or get the list of slices for a user
      operationId: getSlicesUsingGET
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: user
          in: query
          description: user
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/SliceResource'
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - slice-controller
      summary: Create a new Slice on the SDK and a Project on the Public Catalogue
      operationId: createSliceUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - in: body
          name: slice
          description: slice
          required: true
          schema:
            $ref: '#/definitions/SliceResource'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Slice created
          schema:
            type: object
        '400':
          description: Slice already present in database or slice cannot be validated
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: Not Found
        '500':
          description: Slice cannot be created on Public Catalogue
      deprecated: false
  '/sdk/sliceManagement/slices/{sliceId}':
    get:
      tags:
        - slice-controller
      summary: Search a Slice with ID
      operationId: getSliceUsingGET
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: sliceId
          in: path
          description: sliceId
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SliceResource'
        '400':
          description: Query without parameter sliceId
        '401':
          description: Authorization header is missing or format not valid
        '403':
          description: Operation not permitted
        '404':
          description: Slice not found in database
      deprecated: false
    delete:
      tags:
        - slice-controller
      summary: Delete a slice from database
      operationId: deleteSliceUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: sliceId
          in: path
          description: sliceId
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '400':
          description: Query without parameter sliceId
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: Slice not found in database
        '409':
          description: Slice cannot be deleted
        '500':
          description: Slice cannot be deleted on Public Catalogue
      deprecated: false
  '/sdk/sliceManagement/slices/{sliceId}/users/{userName}':
    put:
      tags:
        - slice-controller
      summary: Add user to a slice
      operationId: addUserToSliceUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: sliceId
          in: path
          description: sliceId
          required: true
          type: string
        - name: userName
          in: path
          description: userName
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '400':
          description: Query without parameter sliceId or userName or user already present in the slice
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: Slice not found in database or user not found in Keycloak
      deprecated: false
    delete:
      tags:
        - slice-controller
      summary: Delete user from a slice
      operationId: delUserFromSliceUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: sliceId
          in: path
          description: sliceId
          required: true
          type: string
        - name: userName
          in: path
          description: userName
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '400':
          description: Query without parameter sliceId or userName
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: Slice not found in database or user not found in the slice
      deprecated: false
definitions:
...
  SliceResource:
    type: object
    properties:
      id:
        type: integer
        format: int64
      sliceDescription:
        type: string
      sliceId:
        type: string
      users:
        type: array
        items:
          type: string
    title: SliceResource
...